feat(utils.ts): add isValidUUID function to validate UUID format for user ID refactor(users API): enhance error handling and UUID validation in user-related API endpoints, improving robustness and user feedback chore(api): remove unused test-db and u API files to clean up the codebase

+++ src/pages/api/users/[userId].ts
@@ -1,38 +1,65 @@
+import { isValidUUID } from "@/lib/utils";
 import { deleteUser, getUserById, updateUser } from "@/services/userService";
 import { NextApiRequest, NextApiResponse } from "next";
+
 export default async function handler(
   req: NextApiRequest,
   res: NextApiResponse,
 ) {
   const { userId } = req.query as { userId?: string };
-  console.log("userId", userId);
 
   switch (req.method) {
     case "GET":
       if (userId) {
-        // Get a single user by ID
-        const user = await getUserById(userId);
-        if (user) {
-          res.status(200).json(user);
-        } else {
-          res.status(404).json({ message: "User not found" });
+        if (!isValidUUID(userId)) {
+          res.status(400).json({ message: "Invalid UUID format" });
+          return;
+        }
+
+        try {
+          const user = await getUserById(userId);
+          if (user) {
+            res.status(200).json(user);
+          } else {
+            res.status(404).json({ message: "User not found" });
+          }
+        } catch (error: any) {
+          res
+            .status(500)
+            .json({ message: "Error fetching user", error: error.message });
         }
       }
-
+      break;
     case "PUT": // Update a user by ID
-      if (!userId) {
-        return res.status(400).json({ message: "User ID is required" });
+      if (!userId || !isValidUUID(userId)) {
+        res.status(400).json({ message: "Invalid UUID format" });
+        return;
       }
-      const updatedUser = await updateUser(userId, req.body);
-      res.status(200).json(updatedUser);
+
+      try {
+        const updatedUser = await updateUser(userId, req.body);
+        res.status(200).json(updatedUser);
+      } catch (error: any) {
+        res
+          .status(500)
+          .json({ message: "Error updating user", error: error.message });
+      }
       break;
 
     case "DELETE": // Delete a user by ID
-      if (!userId) {
-        return res.status(400).json({ message: "User ID is required" });
+      if (!userId || !isValidUUID(userId)) {
+        res.status(400).json({ message: "Invalid UUID format" });
+        return;
       }
-      const deletedUser = await deleteUser(userId);
-      res.status(200).json(deletedUser);
+
+      try {
+        const deletedUser = await deleteUser(userId);
+        res.status(200).json(deletedUser);
+      } catch (error: any) {
+        res
+          .status(500)
+          .json({ message: "Error deleting user", error: error.message });
+      }
       break;
 
     default:
